moved quick print script and added image deletion
