--- image-combiner/.python-version
@@ -1,1 +1,0 @@
-venv3-8-10
