--- image-combiner/README.md
@@ -1,14 +1,0 @@
-# Image combiner
-
-Adds the logo to the photo, then duplicates for a 2-up layout
-
-
-## Usage
-```
-python image-combiner path/to/image.jpg
-```
-
-## Dependencies
-```
-python -m pip install --upgrade Pillow
-```
