--- image-combiner/quick-print.scm
@@ -1,7 +1,0 @@
-(define (quick-print inImage)
-  (let* (
-         (image (car(gimp-file-load RUN-NONINTERACTIVE inImage inImage)))
-         )
-    (file-print-gtk RUN-INTERACTIVE image)
-  )
-)
