+++ photo/.python-version
@@ -1,0 +1,1 @@
+venv3-8-10
