+++ photo/images-output/.gitignore
@@ -1,0 +1,3 @@
+*
+*/
+!.gitignore
